+++ test/server/utils/testUser.js
@@ -1,0 +1,72 @@
+const Authenticator = require('./../../../server/authenticator').default;
+
+/**
+ *  Manages a test user for testing purposes.
+ * 
+ *  @property {string} username
+ *  @property {string} password
+ *  @class
+ */
+class TestUser {
+    constructor(options = {}) {
+        this.username = options.username || 'test';
+        this.password = options.password || 'test';
+        this.init();
+    }
+
+    /**
+     *  Invoked at the class initialization.
+     */
+    init() {
+        this.removeUser();
+        this.createUser();
+    }
+
+    /**
+     *  Removes user with the same username if exists.
+     */
+    removeUser() {
+        Meteor.users.remove({
+            username: this.username
+        });
+    }
+
+    /**
+     *  Creates username for the configuration credentials.
+     */
+    createUser() {
+        Accounts.createUser({
+            username: this.username,
+            password: this.password
+        });   
+    }
+
+    /**
+     *  Creates and sets login tokens for the user according
+     *  to the provided to the method options.
+     *  @param {Object} options 
+     *  @returns {boolean} result
+     */
+    setLoginToken({ resume, rememberMe = true }) {
+        const loginToken = {
+            when: new Date(),
+            hashedToken: Authenticator.hashToken(resume),
+            rememberMe
+        };
+        
+        const user = Meteor.users.findOne({ username: this.username });
+        if (!user) {
+            return false;
+        }
+
+        const result = Meteor.users.update(user._id, {
+            $set: {
+                'services.resume.loginTokens': [loginToken]
+            }
+        });
+
+        return result === 1;
+    }
+}
+
+module.exports = TestUser;