+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 
 ## [0.1.0] - 19.01.2018
 * Initial release
+
+## [0.1.1] - 27.01.2018
+* Print correct error in case of already dissallowed attempt
+* Add server side tests